Get apexd to verify manifest data
Send the name and version we read from the manifest to apexd to make
sure it doesn't change.

Remove image hash since we have no use for it.

Put the manifest extraction & verification behind a flag since it now
has the potential to change behavior.

Expand on the comment for verify(), to make it clearer what it does
and doesn't verify.

Bug: 313042092
Test: atest ApexTestCases
Test: atest MicrodroidTests
Test: atest libapexutil_rust.test
Change-Id: Ida6d9e11b3bce5676b744dc945eadb09aa9a822f