libfdt: Simplify locating nodes at the root
As the DTSpec defines the offset of the root as 0, simplify .root*() to
instantiate the FdtNode* directly, without involving libfdt. Note that a
future patch may turn these functions from -> Result<_> to -> _ as they
should never fail.

Use that to locate nodes at hard-coded paths which are subnodes of the
root by calling .root().subnode(), which removes the need for libfdt to
do any path parsing and results in a single (and simple) C call:

    fdt_subnode_offset_namelen(fdt, 0, name, name.len());

which simply parses the root.

Note that the same change can't be made for the mutable versions because
the borrow checker complains about the way we are getting subnodes of
FdtNodeMut instances.

Test: m pvmfw
Test: atest liblibfdt.integration_test
Change-Id: I1bcdc9f7a4883e31ef60cfaed61f26465386c996