pvmfw: Introduce CpuInfo for DT parsing
Refactor the code to add support for per-vCPU DT information. The struct
is currently empty and will be used by future CLs.

Similarly to the way UARTs are handled, discard extra CPU nodes with a
warning instead of aborting the boot.

Check at build time that gic_patched_size() won't panic when unwrapped.

Test: m pvmfw
Change-Id: I5f985c6afca9fc909d12f61857ab12ca9eee0ecf