pvmfw: Refactor get_or_generate_instance_salt
This is a rather complicated method that looks up instance.img, get the
salt if present, generates it if not. Additionally, it matches dice
measurement with recorded information from first boot. It also records
all this information in instance.img. It additionally has logic to skip
the checks for rkpvm.

Refactor this to get_recorded_entry / dice_measurements_matching /
record_instance_entry.

Test: Builds
Bug: 291213394
Change-Id: I6b1e14ae759660b49519f9b64317b9cf74450edd