Add libminijail to the compos APEX
Due to the symlink optimization, the compos APEX didn't have libminijail
in it even though it has a transitive dependency to it.

This is problematic because compos APEX is also loaded in Microdroid and
the existence of libminijail is not guaranteed there.

In order for the APEX to not depend on Microdroid, add the shared lib
explicitly to the APEX.

Bug: 334192594
Test: build com.android.compos and check that libminijail is in it.
Change-Id: I740f08cd42eeac602cf3198f6fd593c3620d5146