Copy MotionEvent before putting it into the queue
A MotionEvent obj can be recycled, and in that case, the native data
parts can be rewritten by a new event before the stored object is
handled in the background thread. If the new event is 'up' event, the
data in previous MotionEvent for the last pointer in 'move' event
becomes invalid.

Bug: 353928688
Test: finger up and down repeatly
Change-Id: Iaef74ab2265bb2697443ad6c97eeff2e4ee95f05