Eliminate a short blackout during the resume of the VM display
Before this CL, there was a short period of blackout (1-2 sec) on the
VM's display when it is brought to foreground. This was because we send
a lid-close even when the VM goes background, and a lid-open event when
it goes foreground, and the guest OS needs some time to respond to the
lid-open event and draw a first frame after the wakeup.

This CL fixes the blackout issue, by saving the last good frame just
before the VM goes background, and re-draws the frame when it goes
foreground.

Note that the saved frame could be very old (ex: a few hours ago), so it
may be incorrect. For example, if the guest OS has a clock, the clock on
the saved frame may point to a very far past. However, this is not a big
issue because the guest OS will soon (within 1-2 sec) react to the
lid-open event and draw a correct frame.

Bug: 348380730
Test: run ferrochrome. go to home, and then come back. There's no
blackout.

Change-Id: I895cb04f3aa32f0b2a65947720e3af417af6296c