[rkpvm] Decide whether VM is protected using FDT property
It is not reliable to use MMIO support to decide whether rialto
is running in protected mode or not, as certain hypervisors
support protected VMs but do not implement MMIO.

This cl replaces this check with an FDT property check.
The property is only set in pvmfw.

We note that this check is only relevant for testing. In
production environments, rialto always operates in protected
mode, rendering it inconsequential if a malicious host attempts
to manipulate the DT by inserting this property.

Bug: 353758405
Test: atest rialto_test
Change-Id: Ic33be0fe1fcd1791515d50e31329d7cb9e873107