rialto: Add test non_protected_vm_with_extra_ram
Export the VM memory size used by service_vm_manager (for its pVM) so
that out non-protected test can use it and remain in-sync.

Introduce a new test using a test-defined size and, for now, use the
same size as the existing tests. As the current size (300MiB) is
consideraby larger than the payload needs, it is expected to be reduced.
This new test will act as a comparison point, to easily identify any
potential breakage caused by the shrinkage.

Name the VM to include the memory size to avoid contention for the
name between the 2 non-protected VMs.

Test: atest rialto_test
Change-Id: Id7d0b9cfa1799dc16fc16e749a8a516a07cd7532