autfs: Replace use of cfg_if with hardened cast
The previous code seemed to guesstimate the C type blksize_t, type of
stat::st_blksize (see stat64(2)), by using Rust config flags so that it
could cast explicitly a u64 constant to it.

Instead, trust that the libc crate got the type width right and cast the
constant to it in a safe way by relying on .try_into().unwrap() instead
of `as`, which removes the risk of unexpected information loss ("modulo
2^(WIDTH) behavior"). Panicking instead of returning an error is fine as
it is deterministic for a given build and very unlikely.

As this was the only user of cfg_if, remove it from the Android.bp.

Note: no behavioral change intended.

Test: banchan com.android.virt aosp_arm64 && m apps_only dist
Change-Id: I6f92fc1a78297b852b26d6fd5f10e3284e8737fe