virtmgr: don't depend on crosvm's libbase_rust module
`base::UnixSeqpacketListener::bind` is just setting up a socket, which
we can easily reproduce without a crosvm dependency.

crosvm sets the backlog to 128, which is the max supported by bionic.
The `nix` crate tries to validate the value isn't beyond the max, but
has an off-by-one error that makes the max value unusable, so I've
lowered it to 127. It is unlikely we'd hit either limit in practice.

Bug: 269356487
Test: boot microdroid
Change-Id: I24560e120374c825e419c16b7c5b5e3d7a5f5b25